requirements.txt
python-telegram-bot==13.14
psycopg2-binary==2.9.6
openai==0.27.8
python-decouple==3.8
